agent: cd to the directory where agent is run from
This commit causes the agent to `cd` to thes directory where the monkey
agent binary is stored. This is particularly desirable when `data_dir`
is set and the agent is being started from the Island, which is an
AppImage. Since AppImages are read-only filesystems, the agent cannot
create any files in `./` unless it has already changed CWD to a writable
directory.